model level & db structure

+++ src/main/java/com/booking/train/ticket/dao/StationDao.java
@@ -15,8 +15,8 @@ public interface StationDao {
 
     void update(Station station);
 
-    Station getStationById(long id);
+    Station getStationById(Integer id);
 
-    void delete(long id);
+    void delete(Integer id);
 
 }
